requirements.txt
numpy==1.19.2
opencv-python==4.5.1.48
pdf2image==1.14.0
Pillow==8.1.0
pytesseract==0.3.7